refactor: modularize commands and handlers (#48)
