+++ src/core/deleteExistingCommands.ts
@@ -1,6 +1,6 @@
 import { REST, Routes } from 'discord.js';
 
-import { config } from './config';
+import { config } from '../config';
 
 export const deleteExistingCommands = async (
   rest: REST,
