+++ src/modules/coolLinksManagement/summarizeCoolPages.ts
@@ -1,7 +1,7 @@
 import { load } from 'cheerio';
 
-import { config } from './config';
-import { resolveCatch } from './helpers/resolve-catch.helper';
+import { config } from '../../config';
+import { resolveCatch } from '../../helpers/resolveCatch.helper';
 // langfrom can't be changed to another language, this result in a translation of the summary that throw an HTTP error because we are in "FREE PLAN"
 const parseBaseUrl = `${config.thirdParties.pageSummarizerBaseUrl}/convert.php?type=expand&lang=en&langfrom=user&url=`;
