+++ src/modules/coolLinksManagement/summarizeCoolVideos.ts
