+++ src/modules/modules.ts
@@ -1,0 +1,11 @@
+import { coolLinksManagement } from './coolLinksManagement/coolLinksManagement.module';
+import { fart } from './fart/fart.module';
+import { patternReplace } from './patternReplace/patternReplace.module';
+import { voiceOnDemand } from './voiceOnDemand/voiceOnDemand.module';
+
+export const modules = {
+  fart,
+  voiceOnDemand,
+  coolLinksManagement,
+  patternReplace,
+};
